intel/aub: Reset surf_state_offset to 0 after emitting a batch buffer
Normally, when i965 emits surface state information to a batch buffer
and aub tracing is enabled, it calls drm_intel_set_surf_offset() to
tell libdrm where the SURFACE_STATE info exists within the batch
buffer.  Libdrm stores this information away in surf_state_offset so
that later, when it is emitting a batch buffer, it can consult the
SURFACE_STATE info in order to generate the appropriate DUMP_BMP
command that will cause the simulator to dump the rendered image to a
file.

This patch causes surf_state_offset to be reset to 0 at the end of
each batch buffer, so that if a later batch buffer doesn't contain any
SURFACE_STATE info (because it contains only blits, for example),
libdrm won't try to generate more DUMP_BMP commands.

Without this patch, libdrm generates bogus DUMP_BMP commands based on
garbage memory.  In the best case, this would cause the simulator to
output empty bitmaps.  In the worst case, it would sometimes cause the
simulator to abort with an "Invalid GTT entry" error.